approach 1/requirements.txt
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
rank-bm25==0.2.2
numpy==1.26.4
rapidfuzz==3.9.6
streamlit==1.39.0

